+++ server/package.json
@@ -59,7 +59,8 @@
 		"method-override": "^3.0.0",
 		"morgan": "^1.10.0",
 		"mysql2": "^2.1.0",
-		"pug": "^3.0.0"
+		"pug": "^3.0.0",
+		"winston-daily-rotate-file": "^4.5.0"
 	},
 	"jest": {
 		"transform": {
