+++ server/src/controllers/index.ts
@@ -1,7 +1,7 @@
-import UserController from './user_controller';
-import CardController from './card_controller';
-import ServiceController from './service_controller';
-import TopicController from './topic_controller';
-import ActivityController from './activity_controller';
+import UserController from './user-controller';
+import CardController from './card-controller';
+import ServiceController from './service-controller';
+import TopicController from './topic-controller';
+import ActivityController from './activity-controller';
 
 export { UserController, CardController, ServiceController, TopicController, ActivityController };
